requirements.txt
fastapi[all]
uvicorn
sqlmodel
fastapi-mail
python-dotenv
starlette-admin
